More updates to goal files.  Modified three test cases. Removed a printf from rdtsc.c that was machine dependent. There were int vs. long problems with bcdefs.h and pointer-levels.c.

+++ tests/kvasir-tests/bc/bcdefs.h
@@ -100,7 +100,7 @@ typedef struct
 
 /* Code addresses. */
 typedef struct {
-      int pc_func;
+      long pc_func;
       int pc_addr;
     } program_counter;
